Add verification of measurement types (#945)
**Context:**
Verification so far checks that the observables are supported, but not
that the measurement type is.

**Description of the Change:**

- We add a check to `validate_observables`, and change its name to
`validate_measurements`, as it now more generally validates that both
the `MeasurementProcess` and the observables on a measurement are valid.

- Adding this verification caused many tests fail with a
`CompilationError`, claiming the measurement is unsupported. This
happened in two cases:
- In cases where the tests were trying to perform `StateMP`
measurements: There was a bug where `ProgramFeatures` records _any_
device from the new device API as being in finite-shot mode, even when
no shots are passed. This caused the
`qjit_capabilities.measurement_processes` (which the transform relies
on) to be incorrect for the new device API and `shots=None`. This PR
updates how `ProgramFeatures` are initialized to fix the bug.
- Once that was resolved, there were still failures, catching incorrect
usage in tests (specifically using state based measurements on a device
with `shots`, using sample based measurements without `shots`, or mixing
state and sample based measurements in a single `qnode`). These tests
have been updated.

**Related Shortcut Stories:**
[[sc-69062](https://app.shortcut.com/xanaduai/story/69062)]

**Note:**
I haven't updated any TOML files. The OQC device and all the Lightning
TOML files already include measurement processes. The CUDA TOML doesn't,
but I'm not sure which MPs it supports, and it's not affected by these
changes yet because it still uses the old device API.