Add seeded execution to the Catalyst runtime (#936)
**Context:**
The test `test_dynamic_one_shot_several_mcms` in
`frontend/test/pytest/test_mid_circuit_measurement.py` was marked as
skipped because it was flaky: #842

After some investigation, it was found out that the test involves a
measurement, which gives random results. The qjit run of the test is
random, but the reference `default.qubit` run is seeded, so sometimes
the results fall outside the acceptable range of `np.allclose`.

To resolve this, we add a seeding mechanism for qjit. 

**Description of the Change:**
Implemented a random seeding infrastructure for qjit.

The top-level qjit decorator can now take in a string argument
`seed=<unsigned 32-bit integer>`. The default value is `None`, which means an
unseeded run.

The seed will be propagated to the runtime `ExecutionContext`, which
then initializes a PRNG (the c++ `std::mt19937` pseudo random number
generator) in the context.
The PRNG is then sent to individual devices.
When performing measurements, the devices draw according to this PRNG.

A proper seed is found and set, to deterministically resolve the flaky
test `test_dynamic_one_shot_several_mcms`.

**Benefits:** 
- `test_dynamic_one_shot_several_mcms` now reliably passes
- User can seed a qjit run


**Related GitHub Issues:** closes #839 

[sc-66696]