Add tests for exponential extrapolation (#953)
**Context:** The last PR
([first](https://github.com/PennyLaneAI/catalyst/pull/806),
[second](https://github.com/PennyLaneAI/pennylane/pull/5972)) to add
exponential extrapolation capabilities to catalyst.

**Description of the Change:** This PR ensures that
`pennylane.transforms.exponential_extrapolate` works inside catalyst's
`mitigate_with_zne` function.

**Benefits:** ZNE has increased functionality.

**Possible Drawbacks:** As part of the testing, I've removed 0.1 and 0.2
from the list of values being used to create different circuits. This is
because exponential extrapolation is not necessarily as stable as
polynomial fitting. For example, when extrapolating near constant values
exponential extrapolation can struggle due to the fact that a linear fit
is first performed to understand the "direction" (or `sign`) of the
exponential. If the slope is positive, the data is flipped to fit
something that looks like $A\mathrm{e}^{-Bx} + C$ where
$B\in\mathbb{R}_{> 0}$.

An example of this happening can be seen here.
```py
>>> exponential_extrapolate([1,2,3], [0.3894, 0.3894183, 0.38941])
-1.0000000000000059e-06
>>> richardson_extrapolate([1,2,3], [0.3894, 0.3894183, 0.38941])
0.3893551000000072
```

If we want to ensure to continue testing values 0.1 and 0.2 for
polynomial extrapolation, we can create separate tests. Let me know what
would be best.

**Related GitHub Issues:**

fixes https://github.com/PennyLaneAI/catalyst/issues/754