[MLIR] Fix async tests after the JAX update (#995)
**Context:** The asynchronous execution of QNodes is implemented via
MLIR's async dialect. Each function call to a function annotated with
the `qnode` attribute will be wrapped in an `async.execute` operation.
Then the standard `async` lowering pipeline will be triggered.

Values returned by `async.execute` are referenced counted by the
AsyncRuntime. When the AsyncRuntime is destroyed, there is an assertion
to guarantee that there are no values that are still alive.

```c++
  ~AsyncRuntime() {
    threadPool.wait(); // wait for the completion of all async tasks
    assert(getNumRefCountedObjects() == 0 &&
           "all ref counted objects must be destroyed");
  }
```

These values that are referenced counted must have their count updated
by the compiler. [This is achieved via the `async.dropref`
operation.](https://mlir.llvm.org/docs/Dialects/AsyncDialect/#asyncruntimedrop_ref-asyncruntimedroprefop)

The algorithm that determines where to place the `async.dropref`
operations is implemented in the `DetectQNode.cpp` files. Each value
returned by the AsyncRuntime is labeled as a source and a special
function called `catalyst__host__rt__unrecoverable_error` is labeled as
a sink. The sources are expected to be in different basic blocks. We
calculate which values are alive in each sink using liveness and insert
a dropRef just before the sink for each source that is alive. Each sink
is removed as a sink after the pattern is applied.

During the migration to the newer version of MLIR. We noticed that the
assertion above was being triggered. After investigating for a little
while, it was found that MLIR implemented identical code folding for
basic blocks was now being executed during simplification.
Simplification happens when using the GreedyRewriter (that's the fold in
applyPatternsAndFold).

Left is after the update right is before the update.

```
  ^bb2:  // pred: ^bb1                                            ^bb2:  // pred: ^bb1
    llvm.call @mlirAsyncRuntimeDropRef(%9, %5) : (!llvm.ptr,  |     llvm.call @mlirAsyncRuntimeDropRef(%7, %2) : (!llvm.ptr, 
    llvm.call @mlirAsyncRuntimeDropRef(%10, %5) : (!llvm.ptr, |     llvm.call @mlirAsyncRuntimeDropRef(%8, %2) : (!llvm.ptr, 
    llvm.return                                                     llvm.return
  ^bb3(%18: !llvm.ptr):  // 2 preds: ^bb0, ^bb1               |   ^bb3:  // pred: ^bb0
    llvm.call @puts(%18) : (!llvm.ptr) -> ()                  |     %16 = llvm.mlir.addressof @assert_msg : !llvm.ptr
                                                              >     llvm.call @puts(%16) : (!llvm.ptr) -> ()
                                                              >     llvm.call @abort() : () -> ()
                                                              >     llvm.unreachable
                                                              >   ^bb4:  // pred: ^bb1
                                                              >     %17 = llvm.mlir.addressof @assert_msg_0 : !llvm.ptr
                                                              >     llvm.call @puts(%17) : (!llvm.ptr) -> ()
    llvm.call @abort() : () -> ()                                   llvm.call @abort() : () -> ()
    llvm.unreachable                                                llvm.unreachable
  }                                                               }
  ```

This change breaks the analysis of placement of dropRefs. Because the basic block with the abort (which will later have the `catalyst__host__rt__unrecoverable_error`) will need to have different calls to `dropRef` depending on which predecessor is called. This is what the code looked like before the update:

```
    llvm.call @mlirAsyncRuntimeAwaitToken(%12) : (!llvm.ptr) -> ()
%14 = llvm.call @mlirAsyncRuntimeIsTokenError(%12) : (!llvm.ptr) -> i1
    %15 = llvm.xor %14, %6  : i1
    llvm.cond_br %15, ^bb1, ^bb3
  ^bb1:  // pred: ^bb0   // First success
    llvm.call @mlirAsyncRuntimeDropRef(%12, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @mlirAsyncRuntimeDropRef(%13, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @mlirAsyncRuntimeAwaitToken(%9) : (!llvm.ptr) -> ()
%16 = llvm.call @mlirAsyncRuntimeIsTokenError(%9) : (!llvm.ptr) -> i1
    %17 = llvm.xor %16, %6  : i1
    llvm.cond_br %17, ^bb2, ^bb4
  ^bb2:  // pred: ^bb1 // Second success
    llvm.call @mlirAsyncRuntimeDropRef(%9, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @mlirAsyncRuntimeDropRef(%10, %5) : (!llvm.ptr, i64) -> ()
    llvm.return
  ^bb3:  // pred: ^bb0 // First error
    llvm.call @puts(%1) : (!llvm.ptr) -> ()
    llvm.call @mlirAsyncRuntimeAwaitToken(%9) : (!llvm.ptr) -> ()
    llvm.call @mlirAsyncRuntimeAwaitToken(%12) : (!llvm.ptr) -> ()
    llvm.call @mlirAsyncRuntimeDropRef(%9, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @mlirAsyncRuntimeDropRef(%12, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @mlirAsyncRuntimeDropRef(%10, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @mlirAsyncRuntimeDropRef(%13, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @__catalyst__host__rt__unrecoverable_error() : () -> ()
    llvm.unreachable
  ^bb4:  // pred: ^bb1 // Second error
    llvm.call @puts(%0) : (!llvm.ptr) -> ()
    llvm.call @mlirAsyncRuntimeAwaitToken(%9) : (!llvm.ptr) -> ()
    llvm.call @mlirAsyncRuntimeDropRef(%9, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @mlirAsyncRuntimeDropRef(%10, %5) : (!llvm.ptr, i64) -> ()
    llvm.call @__catalyst__host__rt__unrecoverable_error() : () -> ()
    llvm.unreachable
```

When the basic block is merged into one, it is not possible to have different dropref depending on the success and failure.


**Description of the Change:**


To prevent identical basic block folding, we can use the greedy rewriter configuration to avoid region simplification:

```
+        GreedyRewriteConfig config;
+        config.strictMode = GreedyRewriteStrictness::ExistingOps;
+        config.enableRegionSimplification = false;
```

We also move the generation of drop-ref analysis to before canonicalization, since canonicalization will run identical basic block code folding.

There might be some other alternative that would allow us to keep basic block code folding and modify the placement of drop refs. Thinking about it :thinking: 

**Benefits:** Tests pass again.

**Possible Drawbacks:** We need to keep the two basic blocks separated between `--convert-async-to-llvm` and `add-exception-handling-pass`. 

**Related GitHub Issues:** Fixes #971

[sc-70684]