Save caches after each dependency step
When building wheels, all depedencies are built by a single job.
If a dependency fails to build, all depencies have to be rebuilt each
run because of the default cache action behaviour.
By splitting the restore and save steps, we can save a cache entry after
each successfully built dependency.