Support keyword arguments for qjit-compiled functions (#1004)
**Context:**

Currently in Catalyst, keyword arguments to functions are not supported
by qjit. for example:

```
@qjit()
def f(x, y):
    return x * y

result = f(3, y=2)
assert result == f(2, 3) 
```
would result in the following exception raised
```
TypeError: fn() missing 2 required positional arguments: 'x' and 'y'
```

The main reason for this is that `kwargs` is no passed to `jit_compile`
function.

**Description of the Change:**
We pass `kwargs` down to `jit_compile` and then to `capture` and finally
to `trace_to_jaxpr` functions and use it to modify the tracers.
We also added support for grad, jacobian, and value_and_grad. 
vmap already worked fine (note that vmap uses jax.vmap when outside qjit
which didn't work with kwargs when I tested). jvp/vjp dib't support
kwargs which is the same behaviour as the jax versions.

**Benefits:**
User can pass keyword arguments when using qjit. 

**Possible Drawbacks:**
`static_argnums` is still not supported when using keyword arguments. To
support this, we need to first support `static_argnames` to be able to
filter static `kwargs`

**Related GitHub Issues:**
[sc-67133]

---------

Co-authored-by: Romain Moyard <rmoyard@gmail.com>