Remove 'validate_device_capabilities' function and uses (#1045)
**Context:** Validating the simulator's TOML file has helped us catch
cases where a capability was added but the update not propagated to the
TOML file. However, this validation is a big crutch and the issue of
synchronicity should be solved more fundamentally. Additionally, the
extra sections in the TOML file do not work well with modifiers like
adjoint/control, and attempting to encode logic (should a gate be
decomposed or turned into a unitary) is not well suited for a static
format

**Description of the Change:** Remove 'validate_device_capabilities'
function and uses

**Benefits:** Progressively stop using obsolete features in favor of a
more robust device capability system.

**Possible Drawbacks:** We might stop catching cases where a capability
was added in the simulator but not in the TOML file.

[sc-67124]