Fix a bug finding the quantum function jaxpr when using quantum primitives with dynamic one-shot  (#1041)
**Context:**
When dealing with lowering of quantum operations, we need to extract the
quantum function jaxpr by finding the function primitive in the circuit
jaxpr which is assumed to be the first element in jaxpr.eqns. However,
this assumption is wrong specifically when dynamic-one-shot is used. In
this case the quantum function is wrapped in another funcion, therefore
we get a crash trying to access the func primitive at jaxpr.eqns[0].

**Description of the Change:**
implemented a function def _get_call_jaxpr(jaxpr) which iterates through
equations in the jaxpr and finds the func primitive.

**Benefits:**
No specific assumption is made on the location of the quantum function
in the jaxpr

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-68756]