Introduce interface for non-unitary gates
The original definition of the `QuantumGate` interfaces only stipulated
the existance and equal number of input and output qubits.
With the addition of adjoint & control handling in the IR, the interface
was extended to include adjoint & control operands.

A new interface `QuantumOperation` is introduced to represent the
original interface without requiring that the instructions can be
inverted or controlled. The existing interface `QuantumGate` is kept
around to represent unitary/invertible&controllable instructions.

Tablegen base classes implementing the interface methods are also
provided, the existing `Gate_Op` is renamed `UnitaryGate_Op`
(implementing `QuantumGate`), and `Gate_Op` is now used to implement
`QuantumOperation`.