Fix StatePrep/BasisState not differentiable with parameter-shift (#1073)
Fixes #1072 which describes state prep gates not being differentiable
with parameter-shift.

The original definition of the `QuantumGate` interface only stipulated
the existence and equal number of input and output qubits. With the
addition of adjoint & control handling in the IR, the interface was
extended to include adjoint & control operands.
A new interface `QuantumOperation` is introduced to represent the
original interface without requiring that the instructions can be
inverted or controlled. The existing interface `QuantumGate` is kept
around to represent unitary/invertible&controllable instructions.
Tablegen base classes implementing the interface methods are also
provided, the existing `Gate_Op` is renamed `UnitaryGate_Op`
(implementing `QuantumGate`), and `Gate_Op` is now used to implement
`QuantumOperation`.

Also works around an additional issue of stemming from the use of
`debug_assert` in the capture of the BasisState operation.

[sc-72378]