Add full support for JAX linear algebra functions invoked from within a `qjit` block (#1097)
**Context:** Certain JAX linear algebra functions from
[`jax.scipy.linalg`](https://jax.readthedocs.io/en/latest/jax.scipy.html#module-jax.scipy.linalg)
occasionally give incorrect results when invoked from within a `qjit`
block, such as `expm`. Furthermore, many of these linear algebra
function are not supported due to missing LAPACK symbols.

**Description of the Change:** This change fixes the issue where the
linear algebra functions invoked from within a `qjit` block occasionally
yield incorrect results and adds full support for the major linear
algebra functions included in
[`jax.scipy.linalg`](https://jax.readthedocs.io/en/latest/jax.scipy.html#module-jax.scipy.linalg),
including, but not limited to:

*
[`jax.scipy.linalg.cholesky`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.cholesky.html)
*
[`jax.scipy.linalg.expm`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.expm.html)
*
[`jax.scipy.linalg.funm`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.funm.html)
*
[`jax.scipy.linalg.hessenberg`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.hessenberg.html)
*
[`jax.scipy.linalg.lu`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.lu.html)
*
[`jax.scipy.linalg.lu_solve`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.lu_solve.html)
*
[`jax.scipy.linalg.polar`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.polar.html)
*
[`jax.scipy.linalg.qr`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.qr.html)
*
[`jax.scipy.linalg.schur`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.schur.html)
*
[`jax.scipy.linalg.solve`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.solve.html)
*
[`jax.scipy.linalg.sqrtm`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.sqrtm.html)
*
[`jax.scipy.linalg.svd`](https://jax.readthedocs.io/en/latest/_autosummary/jax.scipy.linalg.svd.html)

The root cause of the occasionally incorrect results was that the JAX
BLAS and LAPACK wrapper functions defined in
[jaxlib/cpu/lapack_kernels.cc](https://github.com/google/jax/blob/main/jaxlib/cpu/lapack_kernels.cc)
and
[jaxlib/cpu/lapack_kernels_using_lapack.cc](https://github.com/google/jax/blob/main/jaxlib/cpu/lapack_kernels_using_lapack.cc)
are written assuming FORTRAN-style (column-major) matrix layout, and
invoke the FORTRAN BLAS and LAPACK routines. Catalyst on the other hand
currently assumes C-style (row-major) matrix layout. This change
includes modified versions of the JAX wrapper functions that invoke
their respective BLAS and LAPACK routines using the C interfaces.

**Benefits:** Add support for JAX linear algebra functions to Catalyst.

**Possible Drawbacks:** This fix required extensive modifications to the
JAX wrappers to the BLAS and LAPACK routines. If there are upstream
modifications in JAX to these wrappers, we would need to manually update
them in Catalyst for compatibility with future JAX releases. We would
eventually like to remove the need to manually supply these LAPACK
symbols altogether. This is not currently possible, but is being
discussed in JAX Issue
slight improvement on this approach may be to preserve the column-major
matrix layout in Catalyst so that we may use the FORTRAN BLAS and LAPACK
interfaces shipped with JAX more directly.

**Related GitHub Issues:** #753, closes #1071.

[[sc-72974](https://app.shortcut.com/xanaduai/story/72974/manually-add-support-for-the-important-jax-linear-algebra-functions-in-catalyst)]