ZNE local folding (#1006)
**Context:**

Unitary Fund has partnered with Xanadu PennyLane Catalyst to add local
folding functionality to Catalyst. As discussed previously, the
implementation does not yet match the unit test, but we would like to
share the content state of progress, for feedback.

**Description of the Change:**

This is a work-in-progress to add local folding functionality, of all
gates in a quantum circuit, to Catalyst, with a unit test. (Local
folding of random gates in a circuit will follow, after this PR.)

**Benefits:**

Local folding, of single quantum gates, is an alternative method to
global folding, of the entire circuit effectively as a single overall
operator. Certain circuits and quantum computer architectures might
benefit more from either method.

**Possible Drawbacks:**

There are no particular drawbacks to offering an additional option of
local folding, along with global folding.

**Related GitHub Issues:**

https://github.com/PennyLaneAI/catalyst/issues/755

---------

Co-authored-by: Alessandro Cosentino <cosenal@gmail.com>
Co-authored-by: Romain Moyard <rmoyard@gmail.com>