Change `--split-multiple-tapes` into a module pass (#1130)
**Context:**
It was discovered that if `--split-multiple-tapes` were a function pass,
then the outlining for the multitape functions will be done in parallel,
causing their outlined functions to be created (and hence injected into
the parent module) at stochastic times, so the order of the outlined
functions is not guaranteed. This is causing a mlir lit test to be flaky
https://github.com/PennyLaneAI/catalyst/blob/main/mlir/test/Quantum/SplitMultipleTapesTest.mlir#L67.

Moreover, MLIR mandates that `OperationPass` cannot change the parent
operation of the pass's target
https://mlir.llvm.org/docs/PassManagement/#operation-pass. In our case
the target of the pass is the multitape functions, but the outlined
single tape functions are injected into their parent module. This
violates the MLIR standards.

**Description of the Change:**
Change `--split-multiple-tapes` to a module pass.
Also move the pass to `Quantum` dialect.

**Benefits:** 
No race condition during tests; better conforming to MLIR standards.