Treat DefaultExecutionConfig as if it were immutable
The global, currently mutable, object `DefaultExecutionConfig` should
never be modified. Whenever a modification to the execution config needs
to be made, and the baseline config is `DefaultExecutionConfig`, we deep
copy `DefaultExecutionConfig` (which copies its internal dictionaries as
well) and modify the copy. Where no modification is necessary, we
continue to use `DefaultExecutionConfig` as the default argument to
functions that require it to avoid unnecessary copies.