Prefer creating new `qml.devices.ExecutionConfig` objects over using the global `qml.devices.DefaultExecutionConfig` (#1137)
**Context:** We modify the global execution config object
`qml.devices.DefaultExecutionConfig` in a few places. This can lead to
unexpected bugs and test failures, since many of the functions that use
this object expect it to be in a "clean" state that has not been
modified from its original state. There is a discussion happening in
https://github.com/PennyLaneAI/pennylane/pull/6245 to make
`DefaultExecutionConfig` immutable. In the meantime, we should be
treating it as such in Catalyst, both for functional correctness and in
preparation for this potential change. Better yet, we should prefer
creating new `qml.devices.ExecutionConfig` objects over using the global
variable to minimize risk of unintended modifications to it.

**Description of the Change:** For every function call where
`DefaultExecutionConfig` is passed as a default argument, such as

```python
def preprocess(self, execution_config: ExecutionConfig = DefaultExecutionConfig):
    ...
```
replace it with the following pattern:

```python
def preprocess(self, execution_config: Optional[ExecutionConfig] = None):
    if execution_config is None:
        execution_config = ExecutionConfig()
    ...
```

In places where we explicitly create a new `ExecutionConfig` object,
e.g. in `jax_tracer.py::_make_execution_config()`, we can make
modifications to it directly. In places where we have implicitly created
a new `ExecutionConfig` object, but leave the option open to pass in any
`ExecutionConfig` object, we use `dataclasses.replace()` to make a
shallow copy of the config (with a deep copy of any mutable attributes
that need to be modified), and then modify the copy.

This change also partially reverts #1117 to use the original calls in
`test_device_api.py`.

**Benefits:** Avoid unexpected bugs and test failures due to the
`qml.devices.DefaultExecutionConfig` object being in an altered state.

**Possible Drawbacks:** `DefaultExecutionConfig` is still not inherently
immutable, so we need to continue to be vigilant to not modify it.

**Related GitHub Issues:** Closes #1135.