Extend simplification of chained self inverse gates to all Hermitian gates (#1136)
**Context:**
Currently Catalyst support a decorator called ```cancel_inverses```
which finds and cancels out a patterns where there are two consecutive
self inverse gates. However currently, this pass only support Hadamard
gates. This PR will extend the support to all hermitian gates. The full
list of supported gates are as follows:

- One-bit Gates: Hadamard, PauliX, PauliY, PauliZ

- Two-bit Gates: CNOT, CY, CZ, SWAP

- Three-bit Gates: Toffoli

**Description of the Change:**
Generalized the pattern to find two consecutive gates that are Hermitian
and of same kind where the output of the first gate is the input of the
second gate with the same order of qubits.

**Benefits:**
All Hermitian gates are supported by the cancel_inverses decorator

[sc-67659]

---------

Co-authored-by: paul0403 <79805239+paul0403@users.noreply.github.com>