Update observable decomposition transforms in preprocess (#1123)
**Context:**
We want to add to the logic that decides when to decompose observables
to the measurement basis (and whether to also change the measurement
process to samples/counts) in preprocess. Currently we apply
`measurements_to_samples` or `measurements_to_counts` based on what
measurement types the device supports, but these transforms do more than
just convert to samples or counts - they remove all observables so all
measurements are in the computational basis.

Additionally, as part of a previous story in this epic we have a new
diagonalization transform in PennyLane that can diagonalize unsupported
observables from the set `{X, Y, Hadamard}` (and `op_math` operators
composed of these). That transform isn't being applied in preprocess yet
when relevant.

**Description of the Change:**

We update the logic so that `measurements_from_counts`/`_samples` are
used if the device doesn't support _any_ observables, and so that
`diagonalize_measurements` on used on the relevant observables if the
device only supports a _subset_ of {X, Y, Hadmard} (or only supports
Z-basis observables).

**Benefits:**
Support for more combinations of backend capabilities.

---------

Co-authored-by: Romain Moyard <rmoyard@gmail.com>