Convert to samples/counts based on available MPs (#1148)
**Context:**

In #1123 we changed the criteria for applying
`measurements_from_samples` and `measurements_from_counts` to be based
on whether or not there are supported `observables`. Previously the
decision in preprocess had been based on supported measurement processes
instead of supported observables. The reason for this change is that
those transforms don't only change the measurement type - they also go a
step further and diagonalize the observables to the readout basis. This
is potentially a separate functionality from modifying measurement
types.

There was discussion of this being handled by a future transform that is
more specific to selectively transforming unsupported MPs without
changing the observables. However, that isn't implemented yet, and the
QRack plugin in analytic mode only supports Counts and Samples, so its
needed now.

**Description of the Change:**
This PR reintroduces the check for supported measurement types being
limited to a subset of {"Counts", "Samples"} in when adding measurement
transforms to the device TransformProgram.

**Benefits:**
The QRack device keeps working.

**Possible Drawbacks:**
Depending what the backend supports, arguably the TOML file loading is
what should change here - if the observables are only supported in
`analytic` mode, then the `qjit_capabilities` in finite-shots mode
should have an empty list of `observables`, and then this change would
be unnecessary. So maybe I'm not fixing this in the most logical place?
But I'm not sure about that, and I'm sure this will work.

---------

Co-authored-by: Romain Moyard <rmoyard@gmail.com>