Remove 'puts' error message from async lowering (#645)
**Context:** The async lowering will add a puts and a message about
values in an error state. We should remove them as we now return an
exception. Not really a bug, but a user experience issue.

**Description of the Change:** We renamed the
RemoveAbortInsertCallTransform to RemoveAbortAndPutsInsertCallTransform
to avoid creating a new transform with boiler-plate code. We just have
added the methods and functions that allowed us to, first, collect the
blocks including calls to puts, and second, to erase those call
operations.

**Benefits:** avoid having to create a new transform.

**Possible Drawbacks:** 

- Maybe we want to separate aborts logic from and puts logic?
- Maybe we can delete the puts call ops at the very moment we check for
the blocks containing them?

**Related GitHub Issues:** Fixes
https://github.com/PennyLaneAI/catalyst/issues/514

[sc-59514]

---------

Co-authored-by: Erick Ochoa Lopez <erick.ochoalopez@xanadu.ai>
Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>